reintegrated login/out with store

+++ pages/api/users/logout.js
@@ -1,23 +1,29 @@
 import prisma from "../../../utils/prisma/prisma";
-import { authorize } from "../../../utils/auth/tokens";
+import { deleteCookie } from "../../../utils/auth/cookies";
 
 // delete refresh token
-const logout = async (req, res, refreshToken) => {
+const logout = async (req, res) => {
   if (req.method !== "DELETE") return res.status(405).end();
+  // extract refresh token
+  const refreshToken = req.cookies.jid;
+
+  if (!refreshToken) return res.status(401).end();
 
   try {
     // delete refresh token
     const deleted = await prisma.token.delete({
       where: { token: refreshToken },
     });
-    console.log(deleted);
+
+    // deactivate refresh cookie
+    deleteCookie(res, "jid", true);
     return res.status(200).end();
   } catch (error) {
     console.error(error);
     return res.status(500).end();
   } finally {
-    await prisma.status();
+    await prisma.$disconnect();
   }
 };
 
-export default authorize(logout);
+export default logout;
